+++ rrule-afl-fuzz/run_afl_fuzz.sh
@@ -12,9 +12,9 @@ FOLDER_IN="src"
 FOLDER_OUT="out"
 
 # Build
-cargo afl build --features "no-validation-limits"
+cargo afl build --release
 # Fuzz binary
-BIN_PATH="../target/debug/rrule-afl-fuzz"
+BIN_PATH="../target/release/rrule-afl-fuzz"
 # Env variable to set for all instances
 FUZZ_ENV_VAR="AFL_I_DONT_CARE_ABOUT_MISSING_CRASHES=1 AFL_SKIP_CPUFREQ=1 AFL_MAP_SIZE=131072"
 # What command should be used to open new terminal windows