+++ rrule/src/tests/regression.rs
@@ -25,8 +25,8 @@ RRULE:FREQ=MONTHLY;BYDAY=MO,TU,WE,TH,FR;BYSETPOS=-2"
 
 #[test]
 fn issue_61() {
-    let rrule_set: RRuleSet = "DTSTART;TZID=Europe/Berlin:18930401T010000\nRRULE:FREQ=DAILY"
-        .parse()
+    let rrule_set = "DTSTART;TZID=Europe/Berlin:18930401T010000\nRRULE:FREQ=DAILY"
+        .parse::<RRuleSet>()
         .expect("The RRule is not valid");
 
     let (res, _err) = rrule_set.all_with_error(10);