+++ rrule/src/validator/check_limits.rs
@@ -1,40 +1,33 @@
 use crate::core::DateTime;
-use crate::{validator::ValidationError, RRule, Unvalidated};
+use crate::{validator::ValidationError, RRule};
 
 /// Maximum value of `option.interval` when frequency is yearly.
 /// Limit: 10000 years
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_YEARLY_INTERVAL_MAX: u16 = 10_000;
 /// Maximum value of `option.interval` when frequency is monthly.
 /// Limit: about 83 years
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_MONTHLY_INTERVAL_MAX: u16 = 1_000;
 /// Maximum value of `option.interval` when frequency is weekly.
 /// Limit: about 19 years
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_WEEKLY_INTERVAL_MAX: u16 = 1_000;
 /// Maximum value of `option.interval` when frequency is daily.
 /// Limit: about 27 years
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_DAILY_INTERVAL_MAX: u16 = 10_000;
 /// Maximum value of `option.interval` when frequency is hourly.
 /// Limit: about 416 days
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_HOURLY_INTERVAL_MAX: u16 = 10_000;
 /// Maximum value of `option.interval` when frequency is minutely.
 /// Limit: about 7 days
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_MINUTELY_INTERVAL_MAX: u16 = 10_000;
 /// Maximum value of `option.interval` when frequency is secondly.
 /// Limit: about 13 hours
-/// Does not apply when `no-validation-limits` feature flag is set.
-#[allow(dead_code)]
+/// Does not apply when iteration limit flag is false.
 pub(crate) static FREQ_SECONDLY_INTERVAL_MAX: u16 = 50_000;
 
 /// Check (arbitrary) validator limits.
@@ -44,14 +37,7 @@ pub(crate) static FREQ_SECONDLY_INTERVAL_MAX: u16 = 50_000;
 /// This will prevent the creation of iterators that will always fail.
 /// This validation will limit, if not eliminate, large groups of potential panics when using the
 /// iterator.
-///
-/// When `no-validation-limits` feature is set this function will always return `Ok`.
-/// See README.md for more info.
-#[cfg(not(feature = "no-validation-limits"))]
-pub(crate) fn check_limits(
-    rrule: &RRule<Unvalidated>,
-    dt_start: &DateTime,
-) -> Result<(), ValidationError> {
+pub(crate) fn check_limits(rrule: &RRule, dt_start: &DateTime) -> Result<(), ValidationError> {
     use crate::{validator::YEAR_RANGE, Frequency};
     use chrono::Datelike;
 
@@ -83,21 +69,5 @@ pub(crate) fn check_limits(
     }
 
     // All checked, no errors found
-    Ok(())
-}
-
-/// Check validator limits.
-/// It checks all values in the [`RRule`] and makes sure that they are in
-/// a reasonable range. If the function returns `Ok`, no errors where found.
-///
-/// This will prevent the creation of iterators that will always fail.
-/// This validation will limit, if not eliminate, large groups of potential panics when using the
-/// iterator.
-///
-/// When `no-validation-limits` feature is set this function will always return `Ok`.
-/// See README.md for more info.
-#[cfg(feature = "no-validation-limits")]
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn check_limits(_: &RRule<Unvalidated>, _: &DateTime) -> Result<(), ValidationError> {
     Ok(())
 }